requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.7
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.1
MarkupSafe==1.1.1
packaging==20.7
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.9.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.0.1

